implemented the first pass at bmir-radx/radx-project#37
