+++ test/test_ontology_pull.rb
@@ -76,6 +76,32 @@ class TestOntologyPull < TestCase
     assert_equal 2, ont.submissions.length
   end
 
+  def test_remote_pull_parsing_action
+    ontologies = init_ontologies(1, process_submissions: true)
+    ont = LinkedData::Models::Ontology.find(ontologies[0].id).first
+    ont.bring(:submissions) if ont.bring?(:submissions)
+    assert_equal 1, ont.submissions.length
+
+    # add this ontology to submission queue with :remote_pull action enabled
+    parser = NcboCron::Models::OntologySubmissionParser.new
+    actions = NcboCron::Models::OntologySubmissionParser::ACTIONS.dup
+    actions[:remote_pull] = true
+    parser.queue_submission(ont.submissions[0], actions)
+    parser.process_queue_submissions
+
+    # make sure there are now 2 submissions present
+    ont = LinkedData::Models::Ontology.find(ontologies[0].id).first
+    ont.bring(:submissions) if ont.bring?(:submissions)
+    assert_equal 2, ont.submissions.length
+
+    # verify that no new submission is created when the file has not changed
+    parser.queue_submission(ont.submissions[0], actions)
+    parser.process_queue_submissions
+    ont = LinkedData::Models::Ontology.find(ontologies[0].id).first
+    ont.bring(:submissions) if ont.bring?(:submissions)
+    assert_equal 2, ont.submissions.length
+  end
+
   def test_pull_error_notification
     server_port = Random.rand(55000..65535)
 
@@ -164,8 +190,9 @@ class TestOntologyPull < TestCase
 
   private
 
-  def init_ontologies(submission_count)
-    ont_count, acronyms, ontologies = LinkedData::SampleData::Ontology.create_ontologies_and_submissions(ont_count: 1, submission_count: submission_count, process_submission: false)
+  def init_ontologies(submission_count, process_submissions = false)
+    ont_count, acronyms, ontologies = LinkedData::SampleData::Ontology.create_ontologies_and_submissions(
+                            ont_count: 1, submission_count: submission_count, process_submission: process_submissions)
     ontologies[0].bring(:submissions) if ontologies[0].bring?(:submissions)
     ontologies[0].submissions.each do |sub|
       sub.bring_remaining()
